More refactoring -- outsources Delaunay validation routines
